requirements.txt
dj-database-url==0.4.1
Django==1.9.7
django-rest-framework-mongoengine==3.3.0
djangorestframework==3.3.3
djangorestframework-gis==0.10.1
djangorestframework-jsonapi==2.0.1
geojson==1.3.2
gunicorn==19.6.0
inflection==0.3.1
mongoengine==0.10.6
psycopg2==2.6.1
pymongo==3.2.2
whitenoise==3.2
